+++ src/poetry/core/constraints/version/patterns.py
