+++ src/poetry/core/constraints/version/util.py
@@ -2,11 +2,11 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from poetry.core.semver.version_range import VersionRange
+from poetry.core.constraints.version.version_range import VersionRange
 
 
 if TYPE_CHECKING:
-    from poetry.core.semver.version_constraint import VersionConstraint
+    from poetry.core.constraints.version.version_constraint import VersionConstraint
 
 
 def constraint_regions(constraints: list[VersionConstraint]) -> list[VersionRange]:
