+++ src/poetry/core/factory.py
@@ -230,6 +230,9 @@ class Factory:
         from poetry.core.constraints.generic import (
             parse_constraint as parse_generic_constraint,
         )
+        from poetry.core.constraints.version import (
+            parse_constraint as parse_version_constraint,
+        )
         from poetry.core.packages.dependency import Dependency
         from poetry.core.packages.dependency_group import MAIN_GROUP
         from poetry.core.packages.directory_dependency import DirectoryDependency
@@ -237,7 +240,6 @@ class Factory:
         from poetry.core.packages.url_dependency import URLDependency
         from poetry.core.packages.utils.utils import create_nested_marker
         from poetry.core.packages.vcs_dependency import VCSDependency
-        from poetry.core.semver.helpers import parse_constraint
         from poetry.core.version.markers import AnyMarker
         from poetry.core.version.markers import parse_marker
 
@@ -345,7 +347,7 @@ class Factory:
                 marker = marker.intersect(
                     parse_marker(
                         create_nested_marker(
-                            "python_version", parse_constraint(python_versions)
+                            "python_version", parse_version_constraint(python_versions)
                         )
                     )
                 )
