+++ src/poetry/core/masonry/builders/wheel.py
@@ -21,12 +21,12 @@ from typing import TextIO
 from packaging.tags import sys_tags
 
 from poetry.core import __version__
+from poetry.core.constraints.version import parse_constraint
 from poetry.core.masonry.builders.builder import Builder
 from poetry.core.masonry.builders.sdist import SdistBuilder
 from poetry.core.masonry.utils.helpers import escape_name
 from poetry.core.masonry.utils.helpers import normalize_file_permissions
 from poetry.core.masonry.utils.package_include import PackageInclude
-from poetry.core.semver.helpers import parse_constraint
 
 
 if TYPE_CHECKING:
