+++ src/poetry/core/packages/dependency.py
@@ -11,22 +11,22 @@ from typing import Iterable
 from typing import TypeVar
 
 from poetry.core.constraints.generic import parse_constraint as parse_generic_constraint
+from poetry.core.constraints.version import VersionRangeConstraint
+from poetry.core.constraints.version import parse_constraint
 from poetry.core.packages.dependency_group import MAIN_GROUP
 from poetry.core.packages.specification import PackageSpecification
 from poetry.core.packages.utils.utils import contains_group_without_marker
 from poetry.core.packages.utils.utils import create_nested_marker
 from poetry.core.packages.utils.utils import normalize_python_version_markers
-from poetry.core.semver.helpers import parse_constraint
-from poetry.core.semver.version_range_constraint import VersionRangeConstraint
 from poetry.core.version.markers import parse_marker
 
 
 if TYPE_CHECKING:
     from packaging.utils import NormalizedName
 
+    from poetry.core.constraints.version import VersionConstraint
     from poetry.core.packages.directory_dependency import DirectoryDependency
     from poetry.core.packages.file_dependency import FileDependency
-    from poetry.core.semver.version_constraint import VersionConstraint
     from poetry.core.version.markers import BaseMarker
 
     T = TypeVar("T", bound="Dependency")
@@ -165,8 +165,8 @@ class Dependency(PackageSpecification):
 
     @marker.setter
     def marker(self, marker: str | BaseMarker) -> None:
+        from poetry.core.constraints.version import parse_constraint
         from poetry.core.packages.utils.utils import convert_markers
-        from poetry.core.semver.helpers import parse_constraint
         from poetry.core.version.markers import BaseMarker
         from poetry.core.version.markers import parse_marker
 
@@ -228,8 +228,8 @@ class Dependency(PackageSpecification):
 
     @property
     def base_pep_508_name(self) -> str:
-        from poetry.core.semver.version import Version
-        from poetry.core.semver.version_union import VersionUnion
+        from poetry.core.constraints.version import Version
+        from poetry.core.constraints.version import VersionUnion
 
         requirement = self.pretty_name
