+++ src/poetry/core/packages/project_package.py
@@ -3,13 +3,13 @@ from __future__ import annotations
 from typing import TYPE_CHECKING
 from typing import Any
 
-from poetry.core.semver.helpers import parse_constraint
+from poetry.core.constraints.version import parse_constraint
 from poetry.core.version.markers import parse_marker
 
 
 if TYPE_CHECKING:
     from poetry.core.packages.dependency import Dependency
-    from poetry.core.semver.version import Version
+    from poetry.core.constraints.version import Version
 
 from poetry.core.packages.package import Package
 from poetry.core.packages.utils.utils import create_nested_marker
