+++ tests/packages/test_main.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from typing import cast
 
+from poetry.core.constraints.version import Version
 from poetry.core.packages.dependency import Dependency
 from poetry.core.packages.url_dependency import URLDependency
 from poetry.core.packages.vcs_dependency import VCSDependency
-from poetry.core.semver.version import Version
 
 
 def test_dependency_from_pep_508() -> None:
