+++ tests/packages/utils/test_utils.py
@@ -5,11 +5,11 @@ from pathlib import Path
 import pytest
 
 from poetry.core.constraints.generic import parse_constraint as parse_generic_constraint
+from poetry.core.constraints.version import parse_constraint as parse_version_constraint
 from poetry.core.packages.utils.utils import convert_markers
 from poetry.core.packages.utils.utils import create_nested_marker
 from poetry.core.packages.utils.utils import get_python_constraint_from_marker
 from poetry.core.packages.utils.utils import is_python_project
-from poetry.core.semver.helpers import parse_constraint as parse_version_constraint
 from poetry.core.version.markers import parse_marker
